AaltoFloatsEdge/modules/usbcammodule/requirements.txt
azure-iot-device~=2.7.0
opencv-contrib-python
requests
trollius
tornado~=4.5.3
